Speed up building a beam of square section and constant properties
The function ending with _test are made to test the functions which operate on the previous field (i.e. element..._test is made to test section functions).

constant_p functions are made to build very fast a beam to be used in the tests script to check if everything is ok.